Add search and optimize code
