+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './layouts';
 export * from './cards';
 export * from './lists';
+export * from './inputs';
