+++ src/utils/Constants.ts
@@ -1,0 +1,12 @@
+export const PROJECT_INFO = {
+  distributionWeight: 'Distribution Weight',
+  earliesDelivery: 'Earlies delivery',
+  volume: 'Volume',
+  volumeInTons: 'Volume in tons',
+} as const;
+
+export const SEARCH = 'Search' as const;
+
+export const ROUTES = {
+  marketPlace: 'Marketplace',
+} as const;
