--- src/utlis/Constants.ts
@@ -1,10 +1,0 @@
-export const PROJECT_INFO = {
-    distributionWeight: 'Distribution Weight',
-    earliesDelivery: 'Earlies delivery',
-    volume: 'Volume',
-    volumeInTons: 'Volume in tons',
-  } as const;
-
-export const ROUTES = {
-    marketPlace: 'Marketplace',
-} as const;
